Plot function and normalization
